applied patches by Miriam Ruiz made when debianizing the library

+++ tuxcap/hgeparticle/hgeparticle.cpp
@@ -65,7 +65,7 @@ hgeParticleSystem::hgeParticleSystem(const char *filename, DDImage *sprite, floa
         if (bytes != sizeof(hgeParticleSystemInfo))
           return;
 
-        mbAdditiveBlend = ((((unsigned int)info.sprite) >> 16) & 2) == 0;
+        mbAdditiveBlend = ((((unsigned long int)info.sprite) >> 16) & 2) == 0;
 
 	info.sprite = sprite;
